Added Time to recently opened files
